+++ .github/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Update <small>_ July 2022</small>
 
+- fix: missing error embed for wolfram alpha (18/07/2022)
 - chore: add node-fetch dependency (17/07/2022)
 - chore: remove request dependency (17/07/2022)
 - refactor: requests to node-fetch (17/07/2022)